Refactoring gt4py.storage code
