+++ setup.cfg
@@ -6,7 +6,7 @@
 
 [metadata]
 name = gt4py
-version = attr: gt4py.version.__version__
+version = attr: gt4py.__about__.__version__
 author = ETH Zurich
 author_email = gridtools@cscs.ch
 description = Python API to develop performance portable applications for weather and climate
