+++ src/gt4py/__init__.py
@@ -15,3 +15,5 @@
 """Python API to develop performance portable applications for weather and climate."""
 
 from .__about__ import __version__, __version_info__
+
+from . import eve, storage, cartesian
