+++ tests/storage_tests/__init__.py
@@ -11,15 +11,3 @@
 # distribution for a copy of the license or check <https://www.gnu.org/licenses/>.
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
-# flake8: noqa: F401
-
-"""GridTools storages classes."""
-
-
-from .layout import from_name, register
-from .interface import empty, from_array, full, ones, zeros  # noqa: F401
-
-try:
-    from .interface import dace_descriptor  # noqa: F401
-except ImportError:
-    pass
